# Backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EvoLife Analytics Backend",
    "description": "Analytics backend for the EvoLife ecosystem simulator.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/runs": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Create Simulation Run",
        "operationId": "create_simulation_run_api_runs_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulationRunCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "List Simulation Runs",
        "operationId": "list_simulation_runs_api_runs_get",
        "parameters": [
          {
            "name": "experiment_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Experiment Name"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationRunListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/finish": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Finish Simulation Run",
        "operationId": "finish_simulation_run_api_runs__run_id__finish_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulationRunFinish"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Simulation Run",
        "operationId": "get_simulation_run_api_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationRunDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/snapshots": {
      "post": {
        "tags": [
          "snapshots"
        ],
        "summary": "Submit Snapshot",
        "operationId": "submit_snapshot_api_runs__run_id__snapshots_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulationSnapshotCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationSnapshotResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/snapshots/batch": {
      "post": {
        "tags": [
          "snapshots"
        ],
        "summary": "Submit Snapshots Batch",
        "operationId": "submit_snapshots_batch_api_runs__run_id__snapshots_batch_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulationSnapshotBatchCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationSnapshotBatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/creatures": {
      "post": {
        "tags": [
          "creatures"
        ],
        "summary": "Submit Creature Records",
        "operationId": "submit_creature_records_api_runs__run_id__creatures_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatureLifeRecordBatchCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatureLifeRecordBatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/generations": {
      "post": {
        "tags": [
          "generations"
        ],
        "summary": "Submit Generation Summaries",
        "operationId": "submit_generation_summaries_api_runs__run_id__generations_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerationSummaryBatchCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerationSummaryBatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/population-series": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Get Population Time Series",
        "operationId": "get_population_time_series_api_runs__run_id__population_series_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PopulationTimeSeriesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/evolution-series": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Get Evolution Time Series",
        "operationId": "get_evolution_time_series_api_runs__run_id__evolution_series_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "species",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Species"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvolutionTimeSeriesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreatureLifeRecordBatchCreate": {
        "properties": {
          "records": {
            "items": {
              "$ref": "#/components/schemas/CreatureLifeRecordCreate"
            },
            "type": "array",
            "minItems": 1,
            "title": "Records"
          }
        },
        "type": "object",
        "required": [
          "records"
        ],
        "title": "CreatureLifeRecordBatchCreate"
      },
      "CreatureLifeRecordBatchResponse": {
        "properties": {
          "inserted": {
            "type": "integer",
            "title": "Inserted"
          },
          "records": {
            "items": {
              "$ref": "#/components/schemas/CreatureLifeRecordResponse"
            },
            "type": "array",
            "title": "Records"
          }
        },
        "type": "object",
        "required": [
          "inserted",
          "records"
        ],
        "title": "CreatureLifeRecordBatchResponse"
      },
...
      "GenerationSummaryBatchCreate": {
        "properties": {
          "summaries": {
            "items": {
              "$ref": "#/components/schemas/GenerationSummaryCreate"
            },
            "type": "array",
            "minItems": 1,
            "title": "Summaries"
          }
        },
        "type": "object",
        "required": [
          "summaries"
        ],
        "title": "GenerationSummaryBatchCreate"
      },
      "GenerationSummaryBatchResponse": {
        "properties": {
          "inserted": {
            "type": "integer",
            "title": "Inserted"
          },
          "summaries": {
            "items": {
              "$ref": "#/components/schemas/GenerationSummaryResponse"
            },
            "type": "array",
            "title": "Summaries"
          }
        },
        "type": "object",
        "required": [
          "inserted",
          "summaries"
        ],
        "title": "GenerationSummaryBatchResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RunStatus": {
        "type": "string",
        "enum": [
          "pending",
          "running",
          "completed",
          "failed",
          "cancelled"
        ],
        "title": "RunStatus"
      },
      "SimulationRunCreate": {
        "properties": {
          "experiment_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Experiment Name"
          },
          "random_seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Random Seed"
          },
          "configuration": {
            "additionalProperties": true,
            "type": "object",
            "title": "Configuration"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus",
            "default": "running"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "experiment_name"
        ],
        "title": "SimulationRunCreate"
      },
      "SimulationRunDetailResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "experiment_name": {
            "type": "string",
            "title": "Experiment Name"
          },
          "random_seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Random Seed"
          },
          "configuration": {
            "additionalProperties": true,
            "type": "object",
            "title": "Configuration"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "snapshot_count": {
            "type": "integer",
            "title": "Snapshot Count",
            "default": 0
          },
          "creature_record_count": {
            "type": "integer",
            "title": "Creature Record Count",
            "default": 0
          },
          "generation_summary_count": {
            "type": "integer",
            "title": "Generation Summary Count",
            "default": 0
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "run_id",
          "started_at",
          "experiment_name",
          "status"
        ],
        "title": "SimulationRunDetailResponse"
      },
      "SimulationRunFinish": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/RunStatus",
            "default": "completed"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "SimulationRunFinish"
      },
      "SimulationRunListResponse": {
        "properties": {
          "runs": {
            "items": {
              "$ref": "#/components/schemas/SimulationRunResponse"
            },
            "type": "array",
            "title": "Runs"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "runs",
          "total",
          "limit",
          "offset"
        ],
        "title": "SimulationRunListResponse"
      },
      "SimulationRunResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "experiment_name": {
            "type": "string",
            "title": "Experiment Name"
          },
          "random_seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Random Seed"
          },
          "configuration": {
            "additionalProperties": true,
            "type": "object",
            "title": "Configuration"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "run_id",
          "started_at",
          "experiment_name",
          "status"
        ],
        "title": "SimulationRunResponse"
      },
      "SimulationSnapshotBatchCreate": {
        "properties": {
          "snapshots": {
            "items": {
              "$ref": "#/components/schemas/SimulationSnapshotCreate"
            },
            "type": "array",
            "minItems": 1,
            "title": "Snapshots"
          }
        },
        "type": "object",
        "required": [
          "snapshots"
        ],
        "title": "SimulationSnapshotBatchCreate"
      },
      "SimulationSnapshotBatchResponse": {
        "properties": {
          "inserted": {
            "type": "integer",
            "title": "Inserted"
          },
          "snapshots": {
            "items": {
              "$ref": "#/components/schemas/SimulationSnapshotResponse"
            },
            "type": "array",
            "title": "Snapshots"
          }
        },
        "type": "object",
        "required": [
          "inserted",
          "snapshots"
        ],
        "title": "SimulationSnapshotBatchResponse"
      },
      "SimulationSnapshotCreate": {
        "properties": {
          "simulation_time": {
            "type": "number",
            "minimum": 0.0,
            "title": "Simulation Time"
          },
          "herbivore_population": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Herbivore Population"
          },
          "predator_population": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Predator Population"
          },
          "plant_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Plant Count"
          },
          "births": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Births",
            "default": 0
          },
          "deaths": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Deaths",
            "default": 0
          },
          "average_herbivore_speed": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Herbivore Speed"
          },
          "average_predator_speed": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Predator Speed"
          },
          "average_vision": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Vision"
          },
          "average_lifespan": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Lifespan"
          },
          "average_energy": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Energy"
          },
          "extra_metrics": {
            "additionalProperties": true,
            "type": "object",
            "title": "Extra Metrics"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "simulation_time",
          "herbivore_population",
          "predator_population",
          "plant_count"
        ],
        "title": "SimulationSnapshotCreate"
      },
      "SimulationSnapshotResponse": {
        "properties": {
          "simulation_time": {
            "type": "number",
            "minimum": 0.0,
            "title": "Simulation Time"
          },
          "herbivore_population": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Herbivore Population"
          },
          "predator_population": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Predator Population"
          },
          "plant_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Plant Count"
          },
          "births": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Births",
            "default": 0
          },
          "deaths": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Deaths",
            "default": 0
          },
          "average_herbivore_speed": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Herbivore Speed"
          },
          "average_predator_speed": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Predator Speed"
          },
          "average_vision": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Vision"
          },
          "average_lifespan": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Lifespan"
          },
          "average_energy": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Energy"
          },
          "extra_metrics": {
            "additionalProperties": true,
            "type": "object",
            "title": "Extra Metrics"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "simulation_time",
          "herbivore_population",
          "predator_population",
          "plant_count",
          "id",
          "run_id"
        ],
        "title": "SimulationSnapshotResponse"
      },
...
    }
  }
}